refactor browser code dirs
