+++ dcp_prototype/backend/browser/code/rds/browser_orm.py
@@ -13,10 +13,10 @@ from sqlalchemy import (
 from sqlalchemy.orm import relationship, sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
-pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # noqa
+pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../.."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from browser.config.db_config import BrowserDbConfig
+from browser.code.config.db_config import BrowserDbConfig
 
 
 Base = declarative_base()
