+++ dcp_prototype/backend/browser/scripts/create_db.py
@@ -6,8 +6,8 @@ from sqlalchemy import create_engine
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from browser.config.db_config import BrowserDbConfig
-from browser.rds import browser_orm as orm
+from browser.code.config.db_config import BrowserDbConfig
+from browser.code.rds import browser_orm as orm
 
 
 if __name__ == "__main__":
